api_rest/requirements.txt
asgiref==3.7.2
coverage==7.2.7
Django==4.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
inflection==0.5.1
packaging==23.1
psycopg2-binary==2.9.6
PyJWT==2.7.0
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
